+++ src/component/Card/WishListCard.jsx
@@ -1,19 +1,11 @@
 import "./wishListCard.css";
 import "./card.css";
+import "../index";
+import { useWishList } from "../Context/WishListContext/WishListContext";
 
-const WishListCard = ({ wishlist, setWishList }) => {
+const WishListCard = ({ wishlist }) => {
   const { wishlistitem, wishListCount } = wishlist;
-  const removeFromWishListHandler = (wishListCard) => {
-    setWishList((prev) => {
-      return {
-        ...prev,
-        wishListCount: prev.wishListCount - 1,
-        wishlistitem: prev.wishlistitem.filter(
-          (i) => i._id !== wishListCard._id
-        ),
-      };
-    });
-  };
+  const { removeFromWishListHandler } = useWishList();
 
   return (
     <div className='h-100'>
@@ -21,61 +13,67 @@ const WishListCard = ({ wishlist, setWishList }) => {
       <h2 className='cart-page-header'>
         Your Wishlist has {wishListCount} items
       </h2>
-      {wishlistitem.map((item) => {
-        const {
-          _id,
-          name,
-          price,
-          image,
-          discount,
-          priceBeforeDiscount,
-          discPerc,
-          categoryName,
-          ratings,
-          sale,
-          isOutOfStock,
-          newItem = false,
-        } = item;
-        return (
-          <div className='vertical-cards'>
-            <div className='wishlistCard-container'>
-              <div className='card-container bottom-margin'>
-                <div className='parent-positioning'>
-                  <img src={image} alt={name} />
-                  <div className='inside-container '>
-                    <h3>{name}</h3>
-                    <span className='card-content'>{categoryName}</span>
-                    <p className='card-item-cost'>Rs.{price}</p>
-                    {discount && (
-                      <>
-                        <span className='overline-text margin'>
-                          Rs.{priceBeforeDiscount}
+      <div className='vertical-cards'>
+        {wishlistitem.map((item) => {
+          const {
+            _id,
+            name,
+            price,
+            image,
+            discount,
+            priceBeforeDiscount,
+            discPerc,
+            categoryName,
+            ratings,
+            sale,
+            isOutOfStock,
+            newItem = false,
+          } = item;
+          return (
+            <div>
+              <div className='wishlistCard-container  w-100'>
+                <div className=' card-container bottom-margin'>
+                  <div className='parent-positioning'>
+                    <img src={image} alt={name} />
+                    <div className='inside-container '>
+                      <h3>{name}</h3>
+                      <span className='card-content'>{categoryName}</span>
+                      <p>
+                        <span className='card-item-cost margin'>
+                          Rs.{price}
                         </span>
-                        <span className='sale-perct'>{discPerc}% Off</span>
-                      </>
-                    )}
-                    <div className='rating-xs-margin'>
-                      {ratings}
-                      <span className='fa fa-star checked'></span>
-                      <span className='xs-margin'>|</span>5
+                        {discount && (
+                          <>
+                            <span className='overline-text margin'>
+                              Rs.{priceBeforeDiscount}
+                            </span>
+                            <span className='sale-perct'>{discPerc}% Off</span>
+                          </>
+                        )}
+                      </p>
+                      <div className='rating-xs-margin'>
+                        {ratings}
+                        <span className='fa fa-star checked'></span>
+                        <span className='xs-margin'>|</span>5
+                      </div>
+                      <button className='button card-button ecom-card-button'>
+                        Move to Cart
+                      </button>
+                      <span className='badge-overlay'>
+                        {newItem ? "New" : sale ? "Sale" : null}
+                      </span>
+                      <i
+                        onClick={() => removeFromWishListHandler(item)}
+                        className='fa fa-heart-o icon-btn icon-size icon-overlay contained'
+                      ></i>
                     </div>
-                    <button className='button card-button ecom-card-button'>
-                      Move to Cart
-                    </button>
-                    <span className='badge-overlay'>
-                      {newItem ? "New" : sale ? "Sale" : null}
-                    </span>
-                    <i
-                      onClick={() => removeFromWishListHandler(item)}
-                      className='fa fa-heart-o icon-btn icon-size icon-overlay contained'
-                    ></i>
                   </div>
                 </div>
               </div>
             </div>
-          </div>
-        );
-      })}
+          );
+        })}
+      </div>
     </div>
   );
 };
