+++ src/component/Card/wishListCard.css
@@ -1,6 +1,9 @@
 .h-100 {
   min-height: 100vh;
 }
+.w-100 {
+  width: 100%;
+}
 .wishlistCard-container {
   margin: 1.5rem 3rem 1rem 3rem;
 }
