+++ src/component/ProductsFilter/ProductsFilter.jsx
@@ -106,7 +106,7 @@ const ProductsFilter = () => {
           type='range'
           min='1'
           max='5'
-          value='3'
+          // value='3'
         ></input>
       </label>
     </>
